update map floating action
